+++ apps/health_campaign_field_worker_app/lib/models/entities/mdms_master_enums.dart
@@ -47,4 +47,6 @@ enum MasterEnums {
   serviceDefinitions,
   @MappableValue("FIREBASE_CONFIG")
   firebaseConfig,
+  @MappableValue("SEARCH_HOUSEHOLD_FILTERS")
+  searchHouseHoldFilters,
 }