+++ packages/registration_delivery/lib/blocs/beneficiary_registration/beneficiary_registration.dart
@@ -103,6 +103,12 @@ class BeneficiaryRegistrationBloc
           registrationDate: event.registrationDate,
         ));
       },
+      summary: (value) {
+        emit(value.copyWith(
+          householdModel: event.household,
+          registrationDate: event.registrationDate,
+        ));
+      },
     );
   }
 
@@ -194,9 +200,10 @@ class BeneficiaryRegistrationBloc
       summary: (value) async {
         if (event.navigateToSummary) {
           emit(BeneficiaryRegistrationState.create(
-            addressModel: value.householdModel?.address,
-            householdModel: value.householdModel,
-          ));
+              addressModel: value.householdModel?.address,
+              householdModel: value.householdModel,
+              individualModel: value.individualModel,
+              projectBeneficiaryModel: value.projectBeneficiaryModel));
         } else {
           final individual = value.individualModel;
           final household = value.householdModel;
@@ -279,6 +286,8 @@ class BeneficiaryRegistrationBloc
               BeneficiaryRegistrationPersistedState(
                 navigateToRoot: false,
                 householdModel: household,
+                addressModel: address,
+                individualModel: individual,
               ),
             );
           }
@@ -375,6 +384,8 @@ class BeneficiaryRegistrationBloc
             BeneficiaryRegistrationPersistedState(
               navigateToRoot: false,
               householdModel: household,
+              addressModel: address,
+              individualModel: individual,
             ),
           );
         }
@@ -394,13 +405,15 @@ class BeneficiaryRegistrationBloc
         emit(value.copyWith(loading: true));
         try {
           await householdRepository.update(
-            value.householdModel.copyWith(
+            event.household.copyWith(
               clientAuditDetails: ClientAuditDetails(
-                createdBy: value.householdModel.clientAuditDetails!.createdBy,
+                createdBy: value.householdModel.clientAuditDetails?.createdBy ??
+                    value.householdModel.auditDetails!.createdBy.toString(),
                 createdTime:
-                    value.householdModel.clientAuditDetails!.createdTime,
+                    value.householdModel.clientAuditDetails?.createdTime ??
+                        value.householdModel.auditDetails!.createdTime,
                 lastModifiedBy:
-                    value.householdModel.clientAuditDetails!.lastModifiedBy,
+                    RegistrationDeliverySingleton().loggedInUserUuid,
                 lastModifiedTime: DateTime.now().millisecondsSinceEpoch,
               ),
               memberCount: event.household.memberCount,
@@ -421,6 +434,37 @@ class BeneficiaryRegistrationBloc
               await projectBeneficiaryRepository
                   .update(projectBeneficiary.first.copyWith(tag: event.tag));
             }
+          } else {
+            await projectBeneficiaryRepository.create(ProjectBeneficiaryModel(
+                rowVersion: 1,
+                clientReferenceId: IdGen.i.identifier,
+                dateOfRegistration: DateTime.now().millisecondsSinceEpoch,
+                projectId: RegistrationDeliverySingleton().projectId,
+                tenantId: RegistrationDeliverySingleton().tenantId,
+                beneficiaryClientReferenceId:
+                    beneficiaryType == BeneficiaryType.individual
+                        ? value.individualModel.first.clientReferenceId
+                        : value.householdModel.clientReferenceId,
+                clientAuditDetails: ClientAuditDetails(
+                  createdBy: RegistrationDeliverySingleton()
+                      .loggedInUserUuid
+                      .toString(),
+                  createdTime: DateTime.now().millisecondsSinceEpoch,
+                  lastModifiedTime: DateTime.now().millisecondsSinceEpoch,
+                  lastModifiedBy: RegistrationDeliverySingleton()
+                      .loggedInUserUuid
+                      .toString(),
+                ),
+                auditDetails: AuditDetails(
+                  createdBy: RegistrationDeliverySingleton()
+                      .loggedInUserUuid
+                      .toString(),
+                  createdTime: DateTime.now().millisecondsSinceEpoch,
+                  lastModifiedTime: DateTime.now().millisecondsSinceEpoch,
+                  lastModifiedBy: RegistrationDeliverySingleton()
+                      .loggedInUserUuid
+                      .toString(),
+                )));
           }
 
           for (var element in value.individualModel) {
@@ -698,6 +742,12 @@ class BeneficiaryRegistrationState with _$BeneficiaryRegistrationState {
   const factory BeneficiaryRegistrationState.persisted({
     @Default(true) bool navigateToRoot,
     required HouseholdModel householdModel,
+    IndividualModel? individualModel,
+    ProjectBeneficiaryModel? projectBeneficiaryModel,
+    DateTime? registrationDate,
+    AddressModel? addressModel,
+    @Default(false) bool loading,
+    @Default(false) bool isHeadOfHousehold,
   }) = BeneficiaryRegistrationPersistedState;
 
   const factory BeneficiaryRegistrationState.summary({