+++ packages/registration_delivery/lib/pages/search_beneficiary.dart
@@ -6,14 +6,18 @@ import 'package:digit_scanner/pages/qr_scanner.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_keyboard_visibility/flutter_keyboard_visibility.dart';
+import 'package:registration_delivery/blocs/search_households/individual_global_search.dart';
 import 'package:registration_delivery/registration_delivery.dart';
 
 import '../../utils/i18_key_constants.dart' as i18;
+import '../models/entities/status.dart';
 import '../router/registration_delivery_router.gm.dart';
+import '../utils/global_search_parameters.dart';
 import '../utils/utils.dart';
 import '../widgets/back_navigation_help_header.dart';
 import '../widgets/beneficiary/view_beneficiary_card.dart';
 import '../widgets/localized.dart';
+import '../widgets/status_filter/status_filter.dart';
 
 @RoutePage()
 class SearchBeneficiaryPage extends LocalizedStatefulWidget {
@@ -35,6 +39,7 @@ class _SearchBeneficiaryPageState
 
   double lat = 0.0;
   double long = 0.0;
+  List<String> selectedFilters = [];
 
   SearchHouseholdsState searchHouseholdsState = const SearchHouseholdsState(
     loading: false,
@@ -47,7 +52,7 @@ class _SearchBeneficiaryPageState
   void initState() {
     // Initialize the BlocWrapper with instances of SearchHouseholdsBloc, SearchMemberBloc, and ProximitySearchBloc
     blocWrapper = context.read<SearchBlocWrapper>();
-
+    context.read<LocationBloc>().add(const LoadLocationEvent());
     // Listen to state changes
     blocWrapper.stateChanges.listen((state) {
       if (mounted) {
@@ -76,35 +81,40 @@ class _SearchBeneficiaryPageState
           onNotification: (scrollNotification) {
             if (scrollNotification is ScrollUpdateNotification) {
               final metrics = scrollNotification.metrics;
-              if (metrics.atEdge &&
-                  isProximityEnabled &&
-                  searchController.text == '' &&
-                  metrics.pixels != 0) {
-                // [TODO: Handle the null check at Bloc level for Event parameters
-                blocWrapper.proximitySearchBloc
-                    .add(SearchHouseholdsEvent.searchByProximity(
-                  latitude: lat,
-                  longititude: long,
-                  projectId: RegistrationDeliverySingleton().projectId!,
-                  maxRadius: RegistrationDeliverySingleton().maxRadius!,
-                  offset: blocWrapper.proximitySearchBloc.state.offset,
-                  limit: blocWrapper.proximitySearchBloc.state.limit,
-                ));
-              } else if (metrics.atEdge &&
-                  searchController.text != '' &&
-                  metrics.pixels != 0) {
-                blocWrapper.searchByHeadBloc
-                    .add(SearchHouseholdsEvent.searchByHouseholdHead(
-                  searchText: searchController.text,
-                  projectId: RegistrationDeliverySingleton().projectId!,
-                  isProximityEnabled: isProximityEnabled,
-                  offset: blocWrapper.searchByHeadBloc.state.offset,
-                  limit: blocWrapper.searchByHeadBloc.state.limit,
-                ));
+              if (metrics.atEdge && metrics.pixels != 0) {
+                if (RegistrationDeliverySingleton().beneficiaryType ==
+                    BeneficiaryType.individual) {
+                  blocWrapper.individualGlobalSearchBloc
+                      .add(SearchHouseholdsEvent.individualGlobalSearch(
+                          globalSearchParams: GlobalSearchParameters(
+                    isProximityEnabled: isProximityEnabled,
+                    latitude: lat,
+                    longitude: long,
+                    maxRadius: RegistrationDeliverySingleton().maxRadius,
+                    nameSearch: searchController.text,
+                    filter: selectedFilters,
+                    offset:
+                        blocWrapper.individualGlobalSearchBloc.state.offset +
+                            blocWrapper.individualGlobalSearchBloc.state.limit,
+                    limit: blocWrapper.individualGlobalSearchBloc.state.limit,
+                  )));
+                } else {
+                  blocWrapper.houseHoldGlobalSearchBloc
+                      .add(SearchHouseholdsEvent.houseHoldGlobalSearch(
+                          globalSearchParams: GlobalSearchParameters(
+                    isProximityEnabled: isProximityEnabled,
+                    latitude: lat,
+                    longitude: long,
+                    maxRadius: RegistrationDeliverySingleton().maxRadius,
+                    nameSearch: searchController.text,
+                    filter: selectedFilters,
+                    offset: blocWrapper.houseHoldGlobalSearchBloc.state.offset +
+                        blocWrapper.houseHoldGlobalSearchBloc.state.limit,
+                    limit: blocWrapper.houseHoldGlobalSearchBloc.state.limit,
+                  )));
+                }
               }
             }
-            // Return true to allow the notification to continue to be dispatched to further ancestors.
-
             return true;
           },
           child: ScrollableContent(
@@ -138,63 +148,6 @@ class _SearchBeneficiaryPageState
                         builder: (context, locationState) {
                           return Column(
                             children: [
-                              DigitSearchBar(
-                                controller: searchController,
-                                hintText: localizations.translate(
-                                  i18.searchBeneficiary
-                                      .beneficiarySearchHintText,
-                                ),
-                                textCapitalization: TextCapitalization.words,
-                                onChanged: (value) {
-                                  if (value.isEmpty) {
-                                    blocWrapper.clearEvent();
-                                  }
-                                  if (value.trim().length < 2 &&
-                                      !isProximityEnabled) {
-                                    blocWrapper.clearEvent();
-
-                                    return;
-                                  } else if (isProximityEnabled &&
-                                      value.trim().length < 2) {
-                                    blocWrapper.proximitySearchBloc.add(
-                                      SearchHouseholdsEvent.searchByProximity(
-                                        latitude: locationState.latitude!,
-                                        longititude: locationState.longitude!,
-                                        projectId:
-                                            RegistrationDeliverySingleton()
-                                                .projectId!,
-                                        maxRadius:
-                                            RegistrationDeliverySingleton()
-                                                .maxRadius!,
-                                        offset: offset,
-                                        limit: limit,
-                                      ),
-                                    );
-                                  } else {
-                                    blocWrapper.searchHouseholdsBloc.add(
-                                      const SearchHouseholdsClearEvent(),
-                                    );
-                                    blocWrapper.searchByHeadBloc.add(
-                                      SearchHouseholdsEvent
-                                          .searchByHouseholdHead(
-                                        searchText: value,
-                                        projectId:
-                                            RegistrationDeliverySingleton()
-                                                .projectId!,
-                                        isProximityEnabled: isProximityEnabled,
-                                        latitude: locationState.latitude,
-                                        longitude: locationState.longitude,
-                                        maxRadius:
-                                            RegistrationDeliverySingleton()
-                                                .maxRadius,
-                                        tag: null,
-                                        offset: offset,
-                                        limit: limit,
-                                      ),
-                                    );
-                                  }
-                                },
-                              ),
                               locationState.latitude != null
                                   ? Row(
                                       children: [
@@ -218,24 +171,7 @@ class _SearchBeneficiaryPageState
                                                         .maxRadius !=
                                                     null &&
                                                 isProximityEnabled) {
-                                              blocWrapper.proximitySearchBloc
-                                                  .add(
-                                                SearchHouseholdsEvent
-                                                    .searchByProximity(
-                                                  latitude:
-                                                      locationState.latitude!,
-                                                  longititude:
-                                                      locationState.longitude!,
-                                                  projectId:
-                                                      RegistrationDeliverySingleton()
-                                                          .projectId!,
-                                                  maxRadius:
-                                                      RegistrationDeliverySingleton()
-                                                          .maxRadius!,
-                                                  offset: offset,
-                                                  limit: limit,
-                                                ),
-                                              );
+                                              triggerGlobalSearchEvent();
                                             } else {
                                               blocWrapper.clearEvent();
                                             }
@@ -248,6 +184,115 @@ class _SearchBeneficiaryPageState
                                           ),
                                         ),
                                       ],
+                                    )
+                                  : const Offstage(),
+                              DigitSearchBar(
+                                controller: searchController,
+                                hintText: localizations.translate(
+                                  i18.searchBeneficiary
+                                      .beneficiarySearchHintText,
+                                ),
+                                textCapitalization: TextCapitalization.words,
+                                onChanged: (value) {
+                                  if (value.isEmpty) {
+                                    blocWrapper.clearEvent();
+                                  }
+                                  if (value.trim().length < 2 &&
+                                      !isProximityEnabled) {
+                                    blocWrapper.clearEvent();
+
+                                    return;
+                                  } else if (isProximityEnabled &&
+                                      value.trim().length < 2) {
+                                    triggerGlobalSearchEvent();
+                                  } else {
+                                    blocWrapper.searchHouseholdsBloc.add(
+                                      const SearchHouseholdsClearEvent(),
+                                    );
+                                    triggerGlobalSearchEvent();
+                                  }
+                                },
+                              ),
+                              Align(
+                                alignment: Alignment.topLeft,
+                                child: DigitIconButton(
+                                  iconText: getFilterIconNLabel()['label'],
+                                  icon: getFilterIconNLabel()['icon'],
+                                  onPressed: () => showFilterDialog(),
+                                ),
+                              ),
+                              selectedFilters.isNotEmpty
+                                  ? Align(
+                                      alignment: Alignment.topLeft,
+                                      child: SizedBox(
+                                        height:
+                                            MediaQuery.of(context).size.height *
+                                                0.05,
+                                        child: ListView.builder(
+                                            shrinkWrap: true,
+                                            scrollDirection: Axis.horizontal,
+                                            itemCount: selectedFilters.length,
+                                            itemBuilder: (context, index) {
+                                              return Padding(
+                                                  padding: const EdgeInsets.all(
+                                                      kPadding / 2),
+                                                  child: Container(
+                                                    padding:
+                                                        const EdgeInsets.all(
+                                                            kPadding / 2),
+                                                    decoration: BoxDecoration(
+                                                      border: Border.all(
+                                                          color:
+                                                              const DigitColors()
+                                                                  .cloudGray),
+                                                      borderRadius:
+                                                          BorderRadius.circular(
+                                                              kPadding / 2),
+                                                    ),
+                                                    child: Row(
+                                                      children: [
+                                                        Text(
+                                                            localizations.translate(
+                                                                getStatus(
+                                                                    selectedFilters[
+                                                                        index])),
+                                                            style: TextStyle(
+                                                                color: const DigitColors()
+                                                                    .davyGray)),
+                                                        Text(
+                                                            '(${searchHouseholdsState.householdMembers.length})',
+                                                            style: TextStyle(
+                                                                color: const DigitColors()
+                                                                    .davyGray)),
+                                                        const SizedBox(
+                                                            width: kPadding),
+                                                        GestureDetector(
+                                                          onTap: () {
+                                                            setState(() {
+                                                              selectedFilters.remove(
+                                                                  selectedFilters[
+                                                                      index]);
+                                                            });
+                                                            blocWrapper
+                                                                .clearEvent();
+                                                          },
+                                                          child: Container(
+                                                            color:
+                                                                const DigitColors()
+                                                                    .davyGray,
+                                                            child: Icon(
+                                                              Icons.close,
+                                                              color:
+                                                                  const DigitColors()
+                                                                      .white,
+                                                            ),
+                                                          ),
+                                                        )
+                                                      ],
+                                                    ),
+                                                  ));
+                                            }),
+                                      ),
                                     )
                                   : const Offstage(),
                             ],
@@ -299,8 +344,8 @@ class _SearchBeneficiaryPageState
                               long,
                             ),
                             Coordinate(
-                              i.household.address?.latitude,
-                              i.household.address?.longitude,
+                              i.household?.address?.latitude,
+                              i.household?.address?.longitude,
                             ),
                           );
 
@@ -405,5 +450,109 @@ class _SearchBeneficiaryPageState
         ),
       ),
     );
+  }
+
+  getFilterIconNLabel() {
+    return {
+      'label': localizations.translate(
+        i18.searchBeneficiary.filterLabel,
+      ),
+      'icon': Icons.filter_alt
+    };
+  }
+
+  showFilterDialog() async {
+    var filters = await DigitDialog.show(context,
+        options: DigitDialogOptions(
+          titleIcon: Icon(getFilterIconNLabel()['icon'],
+              color: const DigitColors().burningOrange),
+          titleText: getFilterIconNLabel()['label'],
+          content: StatusFilter(selectedFilters: selectedFilters),
+        ));
+
+    if (filters != null) {
+      for (var filter in filters) {
+        if (!selectedFilters.contains(filter)) {
+          setState(() {
+            selectedFilters.add(filter);
+          });
+        }
+      }
+      triggerGlobalSearchEvent();
+    } else {
+      setState(() {
+        selectedFilters = [];
+      });
+      blocWrapper.clearEvent();
+    }
+  }
+
+  triggerClearEvent() {
+    blocWrapper.clearEvent();
+    setState(() {
+      isProximityEnabled = false;
+      searchController.clear();
+      selectedFilters = [];
+    });
+  }
+
+  void triggerGlobalSearchEvent() {
+    blocWrapper.clearEvent();
+    if (RegistrationDeliverySingleton().beneficiaryType ==
+        BeneficiaryType.individual) {
+      blocWrapper.individualGlobalSearchBloc
+          .add(SearchHouseholdsEvent.individualGlobalSearch(
+              globalSearchParams: GlobalSearchParameters(
+        isProximityEnabled: isProximityEnabled,
+        latitude: lat,
+        longitude: long,
+        maxRadius: RegistrationDeliverySingleton().maxRadius,
+        nameSearch: searchController.text,
+        filter: selectedFilters,
+        offset: offset,
+        limit: limit,
+      )));
+    } else {
+      blocWrapper.houseHoldGlobalSearchBloc
+          .add(SearchHouseholdsEvent.houseHoldGlobalSearch(
+              globalSearchParams: GlobalSearchParameters(
+        isProximityEnabled: isProximityEnabled,
+        latitude: lat,
+        longitude: long,
+        maxRadius: RegistrationDeliverySingleton().maxRadius,
+        nameSearch: searchController.text,
+        filter: selectedFilters,
+        offset: offset,
+        limit: limit,
+      )));
+    }
+  }
+
+  String getStatus(String selectedFilter) {
+    final statusMap = {
+      Status.delivered.toValue(): Status.delivered,
+      Status.notDelivered.toValue(): Status.notDelivered,
+      Status.visited.toValue(): Status.visited,
+      Status.notVisited.toValue(): Status.notVisited,
+      Status.beneficiaryRefused.toValue(): Status.beneficiaryRefused,
+      Status.beneficiaryReferred.toValue(): Status.beneficiaryReferred,
+      Status.administeredSuccess.toValue(): Status.administeredSuccess,
+      Status.administeredFailed.toValue(): Status.administeredFailed,
+      Status.inComplete.toValue(): Status.inComplete,
+      Status.toAdminister.toValue(): Status.toAdminister,
+      Status.closed.toValue(): Status.closed,
+      Status.registered.toValue(): Status.registered,
+      Status.notRegistered.toValue(): Status.notRegistered,
+    };
+
+    var mappedStatus = statusMap.entries
+        .where((element) => element.value.name == selectedFilter)
+        .first
+        .key;
+    if (mappedStatus != null) {
+      return mappedStatus;
+    } else {
+      return selectedFilter;
+    }
   }
 }