+++ packages/registration_delivery/lib/pages/summary_page.dart
@@ -5,7 +5,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:registration_delivery/models/entities/additional_fields_type.dart';
 import 'package:registration_delivery/router/registration_delivery_router.gm.dart';
 import 'package:registration_delivery/widgets/back_navigation_help_header.dart';
-import 'package:registration_delivery/widgets/details_card.dart';
+import 'package:digit_components/widgets/atoms/details_card.dart';
 import 'package:registration_delivery/widgets/showcase/showcase_button.dart';
 
 import '../../../widgets/localized.dart';
@@ -23,10 +23,10 @@ class SummaryPage extends LocalizedStatefulWidget {
   });
 
   @override
-  State<SummaryPage> createState() => _SummaryPageState();
+  State<SummaryPage> createState() => SummaryPageState();
 }
 
-class _SummaryPageState extends LocalizedState<SummaryPage> {
+class SummaryPageState extends LocalizedState<SummaryPage> {
   final clickedStatus = ValueNotifier<bool>(false);
 
   String getLocalizedMessage(String code) {
@@ -55,8 +55,6 @@ class _SummaryPageState extends LocalizedState<SummaryPage> {
               if (value.navigateToRoot) {
                 (router.parent() as StackRouter).maybePop();
               } else {
-                print('stack router:');
-                print(router.stack.first);
                 router.popUntil((route) =>
                     route.settings.name == SearchBeneficiaryRoute.name);
                 context.read<SearchBlocWrapper>().searchHouseholdsBloc.add(