+++ packages/registration_delivery/CHANGELOG.md
@@ -46,3 +46,23 @@
 * Added try catch for local repository to handle db lock
 * Localisation support added for missing fields
 * Bug fixes and enhancements
+
+## 1.0.2+1
+
+* Get Precise location on Submitting latitude longitude details
+
+## 0.0.0-dev.8
+
+* Added query builder and registry updated changes
+* Added New screens - house details
+
+## 0.0.0-dev.9
+
+* Updated individual details file for exposing individual state
+* Updated household_overview for BeneficiaryChecklistRoute
+
+## 0.0.0-dev.10
+
+* Added pagination for filter search
+* Updated PageState class to public for customization
+* Updated digit_components